sougou_bot/requirements.txt
selenium>=4.6.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.0
python-dotenv>=0.20.0
